+++ Pixelgraphy/MyPhotosViewController.m
@@ -90,7 +90,7 @@
         PhotoDetailsViewController* controller = (PhotoDetailsViewController*)segue.destinationViewController;
         
         controller.photos = self.dataSource.data;
-        controller.currentIndex = indexPath.row;
+        controller.currentIndex = (int)indexPath.row;
     }
 }
 -(void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath
